medicineapi/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-rest-swagger==2.2.0
django-test-plus==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.15.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
